+++ wemake_python_styleguide/visitors/tokenize/primitives.py
@@ -254,39 +254,3 @@ class WrongStringTokenVisitor(BaseTokenVisitor):
         # but, since we don't recommend `f`-string, this is a low-priority
         modifiers = token.string[:-1]
         self._checker.check_string_modifiers(token, modifiers)
-
-
-@final
-class WrongStringConcatenationVisitor(BaseTokenVisitor):
-    """Checks incorrect string concatenation."""
-
-    _ignored_tokens: ClassVar[frozenset[int]] = frozenset(
-        (
-            tokenize.NL,
-            tokenize.NEWLINE,
-            tokenize.INDENT,
-            tokenize.COMMENT,
-        )
-    )
-
-    def __init__(self, *args, **kwargs) -> None:
-        """Adds extra ``_previous_token`` property."""
-        super().__init__(*args, **kwargs)
-        self._previous_token: tokenize.TokenInfo | None = None
-
-    def visit(self, token: tokenize.TokenInfo) -> None:
-        """Ensures that all string are concatenated as we allow."""
-        self._check_concatenation(token)
-
-    def _check_concatenation(self, token: tokenize.TokenInfo) -> None:
-        if token.exact_type in self._ignored_tokens:
-            return
-
-        if token.exact_type == tokenize.STRING:
-            if self._previous_token:
-                self.add_violation(
-                    consistency.ImplicitStringConcatenationViolation(token),
-                )
-            self._previous_token = token
-        else:
-            self._previous_token = None