+++ CHANGELOG.md
@@ -43,7 +43,9 @@ Semantic versioning in our case means:
 - **Breaking**: Removes `WPS318`, because it is covered by `ruff` formatter
 - **Breaking**: Removes `WPS319`, because it is covered by `ruff` formatter
 - **Breaking**: Removes `WPS320`, because it is covered by `ruff` formatter
+- **Breaking**: Removes `WPS326`, because it is covered by `ruff` linter
 - **Breaking**: Removes `WPS329`, because it is covered by `ruff` linter
+- **Breaking**: Removes `WPS331`, because it is covered by `ruff` linter
 - **Breaking**: Removes `WPS337`, because it is covered by `ruff` formatter
 - **Breaking**: Removes `WPS340`, because it is covered by `ruff` formatter
 - **Breaking**: Removes `WPS341`, because it is covered by `ruff` formatter